Revert "去除提示，用户应该知道自己要干什么"
This reverts commit 23654dc794e527e1ce3eb03c78996bed3d5835d0.

+++ cubian-nandinstall/usr/lib/cubian-nandinstall/install.sh
@@ -243,12 +243,14 @@ if nandPartitionOK;then
         shutdown -h now
     fi
 else
-    umountNand
-    formatNand   
-    echo ""
-    echoRed "*** Reboot is needed! Please re-run cubian-nandinstall after system is up ***"
-    echo ""
-    if promptyn "reboot now?";then
-        shutdown -r now
+    if promptyn "This will completely destory your data on $NAND_DEVICE, Are you sure to continue?"; then
+        umountNand
+        formatNand   
+        echo ""
+        echoRed "*** Plese reboot then run this script again after the system is up ***"
+        echo ""
+        if promptyn "reboot now?";then
+            shutdown -r now
+        fi
     fi
 fi
